+++ src/main/java/com/archisemtle/semtlewebserverspring/vo/ShowProjectApplicantsListResponseVo.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class ShowProjectApplicantsListResponseVo {
 
-    private long totalElements;
+    private int totalElements;
     private int totalPages;
     private int currentPage;
     private List<ApplicantInfo> applicants;
@@ -20,13 +20,13 @@ public class ShowProjectApplicantsListResponseVo {
     @Getter
     @NoArgsConstructor
     public static class ApplicantInfo {
-        private Long applicantId;
+        private Integer applicantId;
         private String name;
         private Date applyDate;
         private String status;
 
         @Builder
-        public ApplicantInfo(Long applicantId, String name, Date applyDate, String status) {
+        public ApplicantInfo(Integer applicantId, String name, Date applyDate, String status) {
             this.applicantId = applicantId;
             this.name = name;
             this.applyDate = applyDate;
@@ -35,7 +35,7 @@ public class ShowProjectApplicantsListResponseVo {
     }
 
     @Builder
-    public ShowProjectApplicantsListResponseVo(long totalElements, int totalPages, int currentPage,
+    public ShowProjectApplicantsListResponseVo(int totalElements, int totalPages, int currentPage,
         List<ApplicantInfo> applicants) {
         this.totalElements = totalElements;
         this.totalPages = totalPages;
